Remove unused brave_shields.mojom.OnFingerprintingBlocked() method
This method is not being called from anywhere as its only caller was
removed in https://github.com/brave/brave-core/commit/0d2a1a21, back
on June 3rd, 2020.

Talking to the security team, this is expected as we no longer report
fingerprinting attempts in shields anymore after moving to farbling,
confirming that this is no longer needed and can be safely removed.